Moved rocket to be a module instead of library
