+++ 3rdparty/modules/rocket/base.h
